Increase version to 0.2 as it deserves it now.
The Windows version has had a few additions since it felt like it was just 0.1, including the thumbnail previewing, saving, and now switching to hqdefault.

See also #3. The Mac version still doesn't have the checkbox to switch to hqdefault, but it might in a few weeks if I get around to it or someone else does.

Also change description to have it say maxres or hq thumbnail rather than just maxresdefault thumbnail.